fix bugs with ssm exec and exec-tagged
