+++ ztictl/cmd/ztictl/main.go
@@ -2,12 +2,10 @@ package main
 
 import (
 	"os"
-
-	"ztictl/internal/cmd"
 )
 
 func main() {
-	if err := cmd.Execute(); err != nil {
+	if err := Execute(); err != nil {
 		os.Exit(1)
 	}
 }
